♻️[Refactor] 폴더 구조 변경 및 정리,  컴포넌트 이름 변경
