--- open-market/src/data/genres.ts
@@ -1,20 +1,0 @@
-const genres: string[] = [
-	"팝",
-	"댄스",
-	"일렉트로닉",
-	"힙합",
-	"R&B",
-	"클래식",
-	"뉴에이지",
-	"락",
-	"발라드",
-	"인디",
-	"재즈/스윙",
-	"라틴",
-	"국악",
-	"월드뮤직",
-	"앰비언트",
-	"트로트",
-	"기타",
-];
-export default genres;
